Commits now follow all file changes

+++ GithubJS/GitJS/file_views.py
@@ -1,9 +1,11 @@
+from django.utils import timezone
+
 from django.http import HttpResponseRedirect, Http404
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required, permission_required
 from django.urls import reverse
 
-from .models import Branch, File
+from .models import Branch, File, Commit
 
 
 @login_required(login_url='login/')
@@ -17,6 +19,8 @@ def single_branch(request, branch_id):
 @permission_required('GitJS.can_edit', raise_exception=True)
 def add_file(request, branch_id):
     branch = get_object_or_404(Branch, id=branch_id)
+    if request.user.username not in branch.project.get_all_participants():
+        return HttpResponseRedirect(reverse("single_branch", args=(branch_id,)))
     if request.method == 'GET':
         return render(request, "file_edit.html", {"branch": branch, "title": "New file", "file_title": "",
                                                   "file_text": "", "form_action": str(branch_id)+"/add_file"})
@@ -39,6 +43,10 @@ def add_file(request, branch_id):
                                                       "form_action": str(branch_id)+"/add_file"})
         except:
             f.save()
+            new_commit_id = Commit.objects.all().order_by('-id')[0].id + 1
+            commit = Commit(id=new_commit_id, branch=branch, committer=request.user.username, date_time=timezone.now())
+            commit.log_message = 'File '+f.title+' added'
+            commit.save()
             return HttpResponseRedirect(reverse("single_branch", args=(branch_id,)))
 
 
@@ -46,6 +54,8 @@ def add_file(request, branch_id):
 @permission_required('GitJS.can_edit', raise_exception=True)
 def edit_file(request, file_id):
     file = get_object_or_404(File, id=file_id)
+    if request.user.username not in file.branch.project.get_all_participants():
+        return HttpResponseRedirect(reverse("single_branch", args=(file.branch.id,)))
     if request.method == 'GET':
         return render(request, "file_edit.html", {"branch": file.branch, "title": "Edit file", "file_title": file.title,
                                                   "file_text": file.text, "form_action": "edit_file/" + str(file_id)})
@@ -66,9 +76,14 @@ def edit_file(request, file_id):
                                                           "file_title": file.title, "file_text": file.text,
                                                           "error_message": error_message,
                                                           "form_action": "edit_file/" + str(file_id)})
+        new_commit_id = Commit.objects.all().order_by('-id')[0].id + 1
+        commit = Commit(id=new_commit_id, branch=file.branch, committer=request.user.username,
+                        date_time=timezone.now())
+        commit.log_message = 'File ' + file.title + ' changed'
         file.title = file_title
         file.text = file_text
         file.save()
+        commit.save()
         return HttpResponseRedirect(reverse("single_branch", args=(file.branch.id,)))
 
 
@@ -76,5 +91,12 @@ def edit_file(request, file_id):
 @permission_required('GitJS.can_edit', raise_exception=True)
 def delete_file(request, file_id):
     file = get_object_or_404(File, id=file_id)
+    if request.user.username not in file.branch.project.get_all_participants():
+        return HttpResponseRedirect(reverse("single_branch", args=(file.branch.id,)))
+    new_commit_id = Commit.objects.all().order_by('-id')[0].id + 1
+    commit = Commit(id=new_commit_id, branch=file.branch, committer=request.user.username,
+                    date_time=timezone.now())
+    commit.log_message = 'File ' + file.title + ' deleted'
     file.delete()
+    commit.save()
     return HttpResponseRedirect(reverse("single_branch", args=(file.branch.id,)))
